Fix unit test
We can mock the HttpHandler function if it's in a separate module from the factory method
