+++ typescript/src/httpHandler.ts
@@ -3,57 +3,11 @@ export interface IHttpHandler {
   get<TOut>(path: string, headers?: { [key: string]: string; }): Promise<TOut>;
 }
 
-function getPriorityBaseUrl(baseUrl: string | undefined): string {
-  if (baseUrl) {
-    return baseUrl;
-  }
-  if (process.env.ORCHESTRATE_BASE_URL) {
-    return process.env.ORCHESTRATE_BASE_URL;
-  }
-  return "https://api.careevolutionapi.com";
-}
-
-function getPriorityApiKey(apiKey: string | undefined): string | undefined {
-  if (apiKey) {
-    return apiKey;
-  }
-  if (process.env.ORCHESTRATE_API_KEY) {
-    return process.env.ORCHESTRATE_API_KEY;
-  }
-  return undefined;
-}
-
-
-export function createHttpHandler(
-  baseUrl: string | undefined,
-  apiKey: string | undefined,
-  additionalHeaders: { [key: string]: string; } | undefined,
-): IHttpHandler {
-  const defaultHeaders = {
-    ...additionalHeaders ?? {},
-    ...{
-      "Content-Type": "application/json",
-      "Accept": "application/json",
-    }
-  } as { [key: string]: string; };
-
-  const priorityApiKey = getPriorityApiKey(apiKey);
-  if (priorityApiKey) {
-    defaultHeaders["x-api-key"] = priorityApiKey;
-  }
-
-  const priorityBaseUrl = getPriorityBaseUrl(baseUrl);
-
-  return new HttpHandler(priorityBaseUrl, defaultHeaders);
-}
 
 export class HttpHandler implements IHttpHandler {
-  baseUrl: string;
-  defaultHeaders: { [key: string]: string; };
-
-  constructor(baseUrl: string, headers: { [key: string]: string; }) {
-    this.baseUrl = baseUrl;
-    this.defaultHeaders = headers;
+  
+  constructor(private baseUrl: string, private defaultHeaders: { [key: string]: string; }) {
+    
   }
 
   private mergeHeaders(headers?: { [key: string]: string; }): { [key: string]: string; } {
