+++ typescript/tests/httpHandler.test.ts
@@ -1,25 +1,22 @@
-// import { describe, it, expect, vi } from 'vitest';
-// import { HttpHandler, createHttpHandler } from '../src/httpHandler';
+import { describe, it, expect, vi } from 'vitest';
+import {HttpHandler} from '../src/httpHandler';
+import { createHttpHandler } from '../src/httpHandlerFactory';
 
-// describe("api base url setup", () => {
-//   vi.mock("../src/httpHandler", () => {
-//     return {
-//       HttpHandler: vi.fn().mockImplementation(() => {
-//         return {
-//           post: vi.fn().mockResolvedValue({}),
-//         };
-//       }),
-//     };
-//   });
-//   // const mockedCreateHttpHandler = createHttpHandler as vi.MockedFunction<typeof createHttpHandler>;
-//   const mockHandler = vi.spyOn(HttpHandler.prototype, "get").mockImplementation((baseUrl: string, headers?: { [key: string]: string; }) => {
-//     return new HttpHandler(baseUrl, headers).get;
-//   });
+vi.mock('../src/httpHandler');
 
-//   it("should use baseUrl from configuration", () => {
-//     const _ = createHttpHandler("argument.api", undefined, {});
-//     // console.log(mockedHttpHandler.mock.calls);
-//     expect(mockHandler).toHaveBeenCalledWith("argument.api", expect.any(Object));
-//     // console.log(mockedCreateHttpHandler);
-//   });
-// });
+describe('createHttpHandler', () => {
+ 
+  it('should create an HttpHandler with the correct baseUrl and apiKey', () => {
+    const baseUrl = 'http://example.com';
+    const apiKey = '12345';
+    
+    createHttpHandler(baseUrl, apiKey, {});
+
+    expect(HttpHandler).toHaveBeenCalledWith(baseUrl, {
+      "Content-Type": "application/json",
+      "Accept": "application/json",
+      "x-api-key": apiKey,
+    });
+  });
+});
+
